+++ acme/services/CSE.py
@@ -247,9 +247,9 @@ def startup(args:argparse.Namespace, **kwargs:Dict[str, Any]) -> bool:
 	# this and other redirect functions to determine the correct file / linenumber
 	# in the log output
 	BackgroundWorkerPool.setLogger(lambda l,m: L.logWithLevel(l, m, stackOffset = 2))
-	BackgroundWorkerPool.setJobBalance(	balanceTarget = Configuration.get('cse.operation.jobBalanceTarget'),
-										balanceLatency = Configuration.get('cse.operation.jobBalanceLatency'),
-										balanceReduceFactor = Configuration.get('cse.operation.jobBalanceReduceFactor'))
+	BackgroundWorkerPool.setJobBalance(	balanceTarget = Configuration.get('cse.operation.jobs.balanceTarget'),
+										balanceLatency = Configuration.get('cse.operation.jobs.balanceLatency'),
+										balanceReduceFactor = Configuration.get('cse.operation.jobs.balanceReduceFactor'))
 
 	console = Console()						# Start the console
 	textUI = TextUI()						# Start the textUI